+++ NEWS.md
@@ -46,6 +46,10 @@ This affects `derive_var_dthcaus()`. (#1727)
 - The `date` field of `event_source()` and `censor_source()` accepts expressions
 now. This affects `derive_param_tte()`. (#1727)
 
+- Argument `keep_vars_source` was added to `derive_extreme_records()` which allows to 
+specify which variables in the new observations should be kept. (#1697) 
+
+
 ## Breaking Changes
 
 - The `aval_fun` argument of `derive_param_exist_flag()` was deprecated in favor